Finished Slide animation, fixed static issue
